backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
supabase==2.3.0
sqlalchemy==2.0.25

# AI Integration
replicate==0.22.0
pillow==10.2.0

# Payment
stripe==8.0.0

# Utils
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
